# foundry_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AcadAssist API",
    "description": "Backend API powering the AcadAssist AI Study Assistant via Microsoft Foundry.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Interact with the AcadAssist Study Assistant",
        "description": "Processes student queries through the ONE AcadAssist Agent on Microsoft Foundry. The agent orchestrates tool selection, retrieves grounded academic source materials, and returns structured responses with citations.",
        "operationId": "chat_endpoint_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tools/summarize_document": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Summarize an academic document",
        "description": "Summarize a specific academic document or textbook chapter in the knowledge base.",
        "operationId": "summarize_document",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeDocumentParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Summarize Document"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionItem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Action category (e.g. quiz_generated, plan_updated)"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Human-readable description of the action"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "description": "Action-specific structured data"
          }
        },
        "type": "object",
        "required": [
          "type",
          "description"
        ],
        "title": "ActionItem",
        "description": "Structured action generated during tool execution (e.g. quiz ready, study plan created)."
      },
      "ChatRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "UUID of the student user"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "Student's prompt or question"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "message"
        ],
        "title": "ChatRequest",
        "description": "Request payload sent from React to POST /api/chat."
      },
      "ChatResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Agent's grounded response to the student"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceItem"
            },
            "type": "array",
            "title": "Sources",
            "description": "Preserved citations from retrieved knowledge"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionItem"
            },
            "type": "array",
            "title": "Actions",
            "description": "Follow-up actions or structured tool results"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatResponse",
        "description": "Response payload returned from POST /api/chat to React."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceItem": {
        "properties": {
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id",
            "description": "UUID of the retrieved chunk"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title",
            "description": "Title of the source document"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename",
            "description": "Filename of the source document"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number",
            "description": "Page number of the reference"
          },
          "slide_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slide Number",
            "description": "Slide number of the reference"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "Relevance or retrieval score"
          },
          "content_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Snippet",
            "description": "Brief excerpt from the document"
          }
        },
        "type": "object",
        "required": [
          "document_title"
        ],
        "title": "SourceItem",
        "description": "Citation or reference to source material used by the Agent."
      },
...
      "SummarizeDocumentParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Summary mode (e.g. concise, detailed, bullet_points)"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "document_id",
          "mode"
        ],
        "title": "SummarizeDocumentParams"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://lqs5ss0s-8000.inc1.devtunnels.ms"
    }
  ]
}